I think this is correct as per the examples given
I actually wrote a lot of code defeated myself then tried the
array_unique function and it appears to give the appropriate result
however I am counting the sequence which I am not sure is the right
approach. Can't determine if it wants the sequence counted or the number
of times  a individual fish was caught.